requirements.txt
python-dotenv
openai
requests
streamlit
